+++ imports/startup/i18n/sv.js
@@ -4,5 +4,11 @@ T9n.map('sv', {
   'Enter newPassword': 'Mata in nytt lösenord',
   'Enter email': 'Mata in e-post',
   'Enter username': 'Mata in användarnamn',
-  'Enter username or email': 'Mata in användarnamn eller e-post'
+  'Enter username or email': 'Mata in användarnamn eller e-post',
+  error: {
+    accounts: {
+      "Invalid email or username": "Ogiltig e-postadress eller användarnamn"
+    }
+  },
+  'or use': 'Eller använd'
 });